+++ examples/complete/main.tf
@@ -21,31 +21,36 @@ resource "azurerm_resource_group" "this" {
   tags = local.tags
 }
 
-resource "azurerm_mssql_server" "secondary" {
-  name                = "sql-${random_id.this.hex}-002"
-  resource_group_name = azurerm_resource_group.this.name
-  location            = "West Europe"
-  version             = "12.0"
+module "log_analytics" {
+  source = "github.com/equinor/terraform-azurerm-log-analytics?ref=v1.4.0"
 
-  azuread_administrator {
-    login_username              = "azureadmasterlogin"
-    object_id                   = data.azurerm_client_config.current.object_id
-    azuread_authentication_only = true
-  }
+  workspace_name      = "log-${random_id.this.hex}"
+  resource_group_name = azurerm_resource_group.this.name
+  location            = azurerm_resource_group.this.location
+}
+
+module "storage" {
+  source = "github.com/equinor/terraform-azurerm-storage?ref=v10.3.0"
+
+  account_name                 = "st${random_id.this.hex}"
+  resource_group_name          = azurerm_resource_group.this.name
+  location                     = azurerm_resource_group.this.location
+  log_analytics_workspace_id   = module.log_analytics.workspace_id
+  shared_access_key_enabled    = true
+  network_rules_default_action = "Allow"
 }
 
 module "sql" {
   # source = "github.com/equinor/terraform-azurerm-sql?ref=v0.0.0"
   source = "../.."
 
-  database_name        = "sqldb-${random_id.this.hex}"
-  server_name          = "sql-${random_id.this.hex}"
-  resource_group_name  = azurerm_resource_group.this.name
-  location             = azurerm_resource_group.this.location
-  storage_account_name = "st${random_id.this.hex}sql"
-  administrator_login  = "masterlogin"
-  sku_name             = "Basic"
-  max_size_gb          = 2
+  server_name                = "sql-${random_id.this.hex}"
+  resource_group_name        = azurerm_resource_group.this.name
+  location                   = azurerm_resource_group.this.location
+  administrator_login        = "masterlogin"
+  log_analytics_workspace_id = module.log_analytics.workspace_id
+  storage_blob_endpoint      = module.storage.blob_endpoint
+  storage_account_access_key = module.storage.primary_access_key
 
   azuread_administrator = {
     login_username = "azureadmasterlogin"
@@ -60,8 +65,21 @@ module "sql" {
     }
   }
 
-  security_alert_policy_email_account_admins           = true
-  security_alert_policy_email_addresses                = []
+  security_alert_policy_email_account_admins = true
+  security_alert_policy_email_addresses      = []
+
+  tags = local.tags
+}
+
+module "database" {
+  # source = "github.com/equinor/terraform-azurerm-sql//modules/database?ref=v0.0.0"
+  source = "../../modules/database"
+
+  name        = "sqldb-${random_id.this.hex}"
+  server_id   = module.sql.server_id
+  sku_name    = "Basic"
+  max_size_gb = 2
+
   short_term_retention_policy_retention_days           = 7
   short_term_retention_policy_backup_interval_in_hours = 12