+++ main.tf
@@ -1,80 +1,113 @@
-module "server" {
-  source = "./modules/server"
+resource "random_password" "this" {
+  length      = 128
+  lower       = true
+  upper       = true
+  numeric     = true
+  special     = true
+  min_lower   = 1
+  min_upper   = 1
+  min_numeric = 1
+  min_special = 1
+}
 
-  server_name           = var.server_name
-  location              = var.location
-  resource_group_name   = var.resource_group_name
-  administrator_login   = var.administrator_login
-  azuread_administrator = var.azuread_administrator
-  identity              = var.identity
-
-  firewall_rules = var.firewall_rules
-
-  security_alert_policy_email_addresses      = var.security_alert_policy_email_addresses
-  security_alert_policy_email_account_admins = var.security_alert_policy_email_account_admins
-
-  storage_account_name   = var.storage_account_name
-  storage_container_name = var.storage_container_name
+resource "azurerm_mssql_server" "this" {
+  name                         = var.server_name
+  location                     = var.location
+  resource_group_name          = var.resource_group_name
+  version                      = "12.0"
+  administrator_login          = var.administrator_login
+  administrator_login_password = random_password.this.result
+  minimum_tls_version          = "1.2"
 
   tags = var.tags
+
+  dynamic "azuread_administrator" {
+    for_each = var.azuread_administrator != null ? [var.azuread_administrator] : []
+
+    content {
+      login_username              = azuread_administrator.value["login_username"]
+      object_id                   = azuread_administrator.value["object_id"]
+      azuread_authentication_only = azuread_administrator.value["azuread_authentication_only"]
+    }
+  }
+
+  lifecycle {
+    ignore_changes = [
+      # Allow admin password to be updated outside of Terraform.
+      administrator_login_password
+    ]
+  }
+
+  dynamic "identity" {
+    for_each = var.identity != null ? [var.identity] : []
+
+    content {
+      type         = identity.value["type"]
+      identity_ids = identity.value["identity_ids"]
+    }
+  }
 }
 
-moved {
-  from = random_password.this
-  to   = module.server.random_password.this
+resource "azurerm_mssql_firewall_rule" "this" {
+  for_each = var.firewall_rules
+
+  name             = each.value.name
+  server_id        = azurerm_mssql_server.this.id
+  start_ip_address = each.value.start_ip_address
+  end_ip_address   = each.value.end_ip_address
 }
 
-moved {
-  from = azurerm_mssql_server.this
-  to   = module.server.azurerm_mssql_server.this
+resource "azurerm_mssql_server_extended_auditing_policy" "this" {
+  server_id              = azurerm_mssql_server.this.id
+  log_monitoring_enabled = true
 }
 
-moved {
-  from = azurerm_mssql_firewall_rule.this
-  to   = module.server.azurerm_mssql_firewall_rule.this
+# Create diagnostic setting for master database to enable server wide.
+resource "azurerm_monitor_diagnostic_setting" "this" {
+  name                       = var.diagnostic_setting_name
+  target_resource_id         = "${azurerm_mssql_server.this.id}/databases/master"
+  log_analytics_workspace_id = var.log_analytics_workspace_id
+
+  dynamic "enabled_log" {
+    for_each = toset(var.diagnostic_setting_enabled_log_categories)
+
+    content {
+      category = enabled_log.value
+    }
+  }
+
+  metric {
+    category = "AllMetrics"
+
+    retention_policy {
+      enabled = false
+    }
+  }
+
+  depends_on = [
+    # Wait for server extended auditing policy to be created.
+    # This ensures the master database exists before trying to create a diagnostic setting for it.
+    azurerm_mssql_server_extended_auditing_policy.this
+  ]
 }
 
-moved {
-  from = azurerm_mssql_server_extended_auditing_policy.this
-  to   = module.server.azurerm_mssql_server_extended_auditing_policy.this
+resource "azurerm_mssql_server_security_alert_policy" "this" {
+  resource_group_name  = azurerm_mssql_server.this.resource_group_name
+  server_name          = azurerm_mssql_server.this.name
+  state                = "Enabled"
+  disabled_alerts      = []
+  email_addresses      = var.security_alert_policy_email_addresses
+  email_account_admins = var.security_alert_policy_email_account_admins
 }
 
-moved {
-  from = azurerm_mssql_server_security_alert_policy.this
-  to   = module.server.azurerm_mssql_server_security_alert_policy.this
-}
+resource "azurerm_mssql_server_vulnerability_assessment" "this" {
+  server_security_alert_policy_id = azurerm_mssql_server_security_alert_policy.this.id
+  storage_container_path          = "${var.storage_blob_endpoint}${var.storage_container_name}/"
+  storage_account_access_key      = var.storage_account_access_key
 
-moved {
-  from = azurerm_storage_account.this
-  to   = module.server.azurerm_storage_account.this
-}
-
-moved {
-  from = azurerm_storage_container.this
-  to   = module.server.azurerm_storage_container.this
-}
-
-moved {
-  from = azurerm_mssql_server_vulnerability_assessment.this
-  to   = module.server.azurerm_mssql_server_vulnerability_assessment.this
-}
-
-module "database" {
-  source = "./modules/database"
-
-  name                 = var.database_name
-  server_id            = module.server.id
-  sku_name             = var.sku_name
-  max_size_gb          = var.max_size_gb
-  storage_account_type = var.database_storage_account_type
-
-  short_term_retention_policy_retention_days           = var.short_term_retention_policy_retention_days
-  short_term_retention_policy_backup_interval_in_hours = var.short_term_retention_policy_backup_interval_in_hours
-
-  long_term_retention_policy_weekly_retention  = var.long_term_retention_policy_weekly_retention
-  long_term_retention_policy_monthly_retention = var.long_term_retention_policy_monthly_retention
-  long_term_retention_policy_yearly_retention  = var.long_term_retention_policy_yearly_retention
-  long_term_retention_policy_week_of_year      = var.long_term_retention_policy_week_of_year
-
-  tags = var.tags
+  recurring_scans {
+    enabled                   = true
+    email_subscription_admins = false
+    emails                    = []
+  }
 }