+++ variables.tf
@@ -1,15 +1,5 @@
-variable "database_name" {
-  description = "The name of this SQL database."
-  type        = string
-}
-
 variable "server_name" {
   description = "The name of this SQL server."
-  type        = string
-}
-
-variable "location" {
-  description = "The location to create the resources in."
   type        = string
 }
 
@@ -18,14 +8,30 @@ variable "resource_group_name" {
   type        = string
 }
 
-variable "storage_account_name" {
-  description = "The name of this Storage account."
+variable "location" {
+  description = "The location to create the resources in."
   type        = string
 }
 
 variable "administrator_login" {
   description = "The login username of the administrator of this SQL server."
   type        = string
+}
+
+variable "log_analytics_workspace_id" {
+  description = "The ID of the Log Analytics workspace to send diagnostics to."
+  type        = string
+}
+
+variable "storage_blob_endpoint" {
+  description = "The blob endpoint Storage account to use for this SQL server."
+  type        = string
+}
+
+variable "storage_account_access_key" {
+  description = "The shared access key of the Storage account to use for this SQL server."
+  type        = string
+  sensitive   = true
 }
 
 variable "azuread_administrator" {
@@ -69,6 +75,18 @@ variable "firewall_rules" {
   }
 }
 
+variable "diagnostic_setting_name" {
+  description = "The name of this diagnostic setting."
+  type        = string
+  default     = "audit-logs"
+}
+
+variable "diagnostic_setting_enabled_log_categories" {
+  description = "A list of log categories to be enabled for this diagnostic setting."
+  type        = list(string)
+  default     = ["SQLSecurityAuditEvents"]
+}
+
 variable "security_alert_policy_email_addresses" {
   description = "A list of custom email addresses to which the alert for this SQL server is sent."
   type        = list(string)
@@ -80,64 +98,11 @@ variable "security_alert_policy_email_account_admins" {
   type        = bool
   default     = true
 }
+
 variable "storage_container_name" {
   description = "The name of this Storage Container."
   type        = string
   default     = "vulnerability-assessment"
-}
-
-variable "sku_name" {
-  description = "The SKU name of this SQL database."
-  type        = string
-  default     = "Basic"
-}
-
-variable "max_size_gb" {
-  description = "The max size of this SQL database in gigabytes."
-  type        = number
-  default     = null
-}
-
-variable "database_storage_account_type" {
-  description = "The type of Storage account used to store backups for this SQL database."
-  type        = string
-  default     = "Geo"
-}
-
-variable "short_term_retention_policy_retention_days" {
-  description = "The number of days that point-in-time restore backups should be retained."
-  type        = number
-  default     = 7
-}
-
-variable "short_term_retention_policy_backup_interval_in_hours" {
-  description = "The hours between each differential backup. Value has to be 12 or 24."
-  type        = number
-  default     = 12
-}
-
-variable "long_term_retention_policy_weekly_retention" {
-  description = "The duration that weekly long-term backups should be retained. Value must be in an ISO 8601 duration format, e.g. `P1Y`, `P1M`, `P1W` or `P7D`."
-  type        = string
-  default     = "PT0S"
-}
-
-variable "long_term_retention_policy_monthly_retention" {
-  description = "The duration that monthly long-term backups should be retained. Value must be in an ISO 8601 duration format, e.g. `P1Y`, `P1M`, `P4W` or `P30D`."
-  type        = string
-  default     = "PT0S"
-}
-
-variable "long_term_retention_policy_yearly_retention" {
-  description = "The duration that yearly long-term backups should be retained. Value must be in an ISO 8601 duration format, e.g. `P1Y`, `P12M`, `P52W` or `P365D`"
-  type        = string
-  default     = "PT0S"
-}
-
-variable "long_term_retention_policy_week_of_year" {
-  description = "The week of year to take the yearly long-term backup. Value must be between `1` and `52`."
-  type        = number
-  default     = 1
 }
 
 variable "tags" {